+++ R/zzz_imports.R
@@ -1,4 +1,5 @@
 # Centralized roxygen-driven imports.
 
 #' @importFrom rlang .data
+#' @importFrom utils head
 NULL